Incorporate built @bazel/typescript; don't hardcode target to es2015
Theia's JSON RPC proxy code fails unless the typescript compiler targets ES5.
The standard bazel typescript rules unfortunately hardcode the target used in
its generated tsconfig files, so we need to fork the library in order to have
control over those settings. Making that fork was a little complicated as rules_nodejs
builds a separate NPM package for the typescript logic; that built packge is included
in this fork so everything is in one place.

This does mean that changes under packages/typescript (the source of the npm package) also
need to be reflected under npm_bazel_typescript. That mainly will come into into play as
we incorporate upstream changes.